move upgrade utils and constants to upgrade_base
